Revert "chore: drop support for SDK 21"
This reverts commit 8f17ce041f4ae8baeb101e0eb0ad636c310a229d.

+++ app/src/main/java/io/github/qauxv/startup/StartupHook.java
@@ -99,7 +99,9 @@ public class StartupHook {
 
     static void deleteDirIfNecessaryNoThrow(Context ctx) {
         try {
-            deleteFile(new File(ctx.getDataDir(), "app_qqprotect"));
+            if (Build.VERSION.SDK_INT >= 24) {
+                deleteFile(new File(ctx.getDataDir(), "app_qqprotect"));
+            }
             if (new File(ctx.getFilesDir(), "qn_disable_hot_patch").exists()) {
                 deleteFile(ctx.getFileStreamPath("hotpatch"));
             }
